+++ user-licence-lifecycle-manager/src/main/java/de/intension/authentication/authenticators/jpa/LicenceJpaSpi.java
@@ -1,0 +1,33 @@
+package de.intension.authentication.authenticators.jpa;
+
+import de.intension.authentication.authenticators.jpa.LicenceJpaProvider;
+import de.intension.authentication.authenticators.jpa.LicenceJpaProviderFactory;
+import org.keycloak.provider.Spi;
+
+/**
+ * Custom Spi so that the {@link LicenceJpaProvider} can be found
+ * as a provider for persisting licences, finding persisted licences and deleting licences from the keycloak database.
+ */
+public class LicenceJpaSpi implements Spi {
+    private static final String LICENCE_JPA = "licenceJpa";
+
+    @Override
+    public boolean isInternal() {
+        return false;
+    }
+
+    @Override
+    public String getName() {
+        return LICENCE_JPA;
+    }
+
+    @Override
+    public Class<LicenceJpaProvider> getProviderClass() {
+        return LicenceJpaProvider.class;
+    }
+
+    @Override
+    public Class<LicenceJpaProviderFactory> getProviderFactoryClass() {
+        return LicenceJpaProviderFactory.class;
+    }
+}