+++ user-licence-lifecycle-manager/src/main/java/de/intension/spi/RestClientProvider.java
@@ -1,0 +1,12 @@
+package de.intension.spi;
+
+import de.intension.rest.LicenceConnectRestClient;
+import org.keycloak.provider.Provider;
+
+public interface RestClientProvider
+    extends Provider
+{
+
+    LicenceConnectRestClient restClient();
+
+}